pyproject.toml
[project]
name = "fastapi-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "fastapi>=0.129.0",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.1",
    "logfire>=4.23.0",
    "mistralai>=1.12.2",
    "pydantic-ai>=1.58.0",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.7.0",
    "python-multipart>=0.0.22",
    "reportlab>=4.4.10",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]
